requirements.txt
requests_ntlm
elastic_enterprise_search
pyyaml